tests verify_package_files: make it clearer how to regenerate data
This adds an argument to the nox session, updates the documentation and
adds a help message regarding how to regenerate the test fixtures. I
decided to copy the colorlog function I wrote for another project,
because nox's output is very colorful and the help message otherwise got
lost in a sea of text.