python_metadata: remove License:: classifiers
`License::` classifiers have been deprecated in favor of the new PEP 639
`license` field that accepts an SPDX expression in pyproject.toml.
setuptools now prints a large deprecation warning when using the
`License::` classifiers. We can't adopt the new standard, as we don't
store configuration in pyproject.toml and do not maintain a cumulative
SPDX license expression that encompasses the additional GPL-compatible
licenses present in some collections. For now, let's just remove
classifier. We still provide a COPYING file for ansible as a whole and
preserve individual collection's license and copyright statements.

```
SetuptoolsDeprecationWarning: License classifiers are deprecated.
!!

        ********************************************************************************
        Please consider removing the following classifiers in favor of a SPDX license expression:

        License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)

        See https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license for details.
        ********************************************************************************

!!
  self._finalize_license_expression()

```

Fixes: https://github.com/ansible-community/antsibull-build/issues/651