tests verify_package_files: make it clearer how to regenerate data (#657)
* tests verify_package_files: make it clearer how to regenerate data

This adds an argument to the nox session, updates the documentation and
adds a help message regarding how to regenerate the test fixtures. I
decided to copy the colorlog function I wrote for another project,
because nox's output is very colorful and the help message otherwise got
lost in a sea of text.

* tests verify_package_files: fix incorrect types

* tests verify_package_files: fix tarfile DeprecationWarning

* tests verify_package_files: remove unused constants

* tests: remove unused tempfile import